pyproject.toml
[project]
name = "cartridge-sdk"
version = "0.1.0"
description = "Cartridge - App SDK for Linux handheld devices"
requires-python = ">=3.9"
dependencies = [
    "pygame-ce>=2.4",
    "aiohttp>=3.9",
    "readability-lxml>=0.8",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
stock = [
    "yfinance>=0.2.31",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
cartridge = "cartridge_sdk.cli.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
